feat(hypervexing): workspace fixes — positions pipeline, bounded HL reads, native registers, stable chart
- fix the proj_open_positions upsert whose CASE-parameter type inference
  failed PREPARE (no perps position was ever projected — empty POSITIONS
  panel); lenient venue markPx parse unblocks hyperliquid.perp.positions
- consolidation chat notice fires only on protection-state transitions;
  the safety wake path still runs on every reconciler pass
- perp/spot markets gain query+limit with a compact single-snapshot
  projection (canonical model-facing decimals, onlyIsolated flag);
  fills/funding gain startTime/limit caps; shared returnedCount/
  matchedCount/truncated envelope; bounded redacted venueError replaces
  opaque batch_error
- five workspace registers render native venue data (TWAP labelled as
  history), positions row shows used leverage; duplicate per-row chart
  and CONSOLIDATING pills removed; risk-setup steppers are click-to-type
  with decimal-preserving clamp; info copy moved to the help modal;
  10-USDC min-notional hint
- chart lifecycle split: chart created once per host, time-range viewport
  preservation with live-edge follow, semantic candle revision (fetchedAt
  inert), imperative price-line handles, manual market pick never
  overridden by auto-follow; compact header typography
- venue-routing prompt: Robinhood stale-reserve caution (negative
  priceImpact / 'Return amount is not enough' => switch to uniswap
  fallback, no slippage retries)

+++ src/vex-agent/db/repos/open-positions.ts
@@ -49,11 +49,20 @@ export interface UpsertPositionRow {
   status?: string;
 }
 
-/** Upsert position — ON CONFLICT updates status, capture-derived MTM, and data. */
+/**
+ * Upsert position — ON CONFLICT updates status, capture-derived MTM, and data.
+ *
+ * $9/$10 (current_value_usd / unrealized_pnl_usd, both NUMERIC) are cast to
+ * ::numeric inside every `CASE WHEN $n IS NULL` use. Without the cast Postgres
+ * cannot infer the parameter type from an `IS NULL` test alone, so PREPARE
+ * failed with `could not determine data type of parameter $9` and NO perps
+ * projection row was ever written — the empty POSITIONS panel in the live app
+ * (incident 2026-07-13). The cast pins the type to the real column type.
+ */
 export async function upsertPosition(row: UpsertPositionRow): Promise<void> {
   await execute(
     `INSERT INTO proj_open_positions (namespace, position_type, chain, external_id, wallet_address, instrument_key, position_key, entry_price_usd, current_value_usd, unrealized_pnl_usd, notional_usd, fee_usd, contracts, settlement_asset_key, data, status, opened_at, last_refresh_at)
-     VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15::jsonb, $16, NOW(), CASE WHEN $9 IS NULL AND $10 IS NULL THEN NULL ELSE NOW() END)
+     VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12, $13, $14, $15::jsonb, $16, NOW(), CASE WHEN $9::numeric IS NULL AND $10::numeric IS NULL THEN NULL ELSE NOW() END)
      ON CONFLICT (namespace, position_type, chain, wallet_address, external_id) WHERE external_id IS NOT NULL
      DO UPDATE SET status = COALESCE($16, proj_open_positions.status),
        data = COALESCE($15::jsonb, proj_open_positions.data),
@@ -66,7 +75,7 @@ export async function upsertPosition(row: UpsertPositionRow): Promise<void> {
        fee_usd = COALESCE($12, proj_open_positions.fee_usd),
        contracts = COALESCE($13, proj_open_positions.contracts),
        settlement_asset_key = COALESCE($14, proj_open_positions.settlement_asset_key),
-       last_refresh_at = CASE WHEN $9 IS NULL AND $10 IS NULL THEN proj_open_positions.last_refresh_at ELSE NOW() END,
+       last_refresh_at = CASE WHEN $9::numeric IS NULL AND $10::numeric IS NULL THEN proj_open_positions.last_refresh_at ELSE NOW() END,
        synced_at = NOW()`,
     [row.namespace, row.positionType, row.chain, row.externalId, row.walletAddress,
      row.instrumentKey ?? null, row.positionKey ?? null, row.entryPriceUsd ?? null,
